server/pyproject.toml
[project]
name = "jarvis-agent"
version = "0.1.0"
description = "Voice AI bot built with Pipecat"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[local-smart-turn-v3,qwen,runner,silero,webrtc,mcp,azure]",
    "pipecat-ai-whisker",
    "pipecat-ai-tail",
    "dashscope>=1.25.5",
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]
